requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==2.3.7

# Scientific Computing
numpy==1.24.3
scipy==1.11.1
matplotlib==3.7.2
plotly==5.15.0

# Quantum Computing
qiskit==0.43.1
qiskit-aer==0.12.0
networkx==3.1

# Finite Element Analysis
sfepy==2023.2

# Parallel Computing & Performance
jax==0.4.13
jaxlib==0.4.13
numba==0.57.1
joblib==1.3.1

# File Processing
vtk==9.2.6
meshio==5.3.4
h5py==3.9.0

# Utilities
python-dotenv==1.0.0
click==8.1.6
redis==4.6.0
celery==5.3.1

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
pre-commit==3.3.3

# Production
gunicorn==21.2.0
supervisor==4.2.5

# Optional: GPU acceleration (uncomment if using CUDA)
# cupy-cuda12x==12.2.0
# cuquantum==23.6.0